Refactored [PossibleMovesIndex] int row,column into byte since index range only 0-7

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/models/PossibleMovesIndex.java
@@ -1,7 +1,7 @@
 package com.thinking.machines.chess.server.models;
 public class PossibleMovesIndex
 {
-public int row;
-public int column;
+public byte row;
+public byte column;
 public boolean safe;
 }
